Add timeout option for requests
Based on an old PR at https://github.com/cjoudrey/gluahttp/pull/20 that
seems to have been deleted.

The timeout option can be set to a number, which will be cast to an int
and interpreted as seconds, or a string which will be parsed by time.ParseDuration.

Add tests, one case should timeout and one that shouldn't.